Update gpio connections

+++ inc/bme280temperature.h
@@ -1,7 +1,7 @@
 #ifndef BME280_TEMPERATURE_H
 #define BME280_TEMPERATURE_H
 
-float bme280_temperature(int *temperature, int *humidity);
+float bme280_temperature(float *temperature, float *humidity);
 void bme280_setup();
 void close_bme280();
